% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.lm_robust.R
\name{anova.lm_robust}
\alias{anova.lm_robust}
\title{anova (model comparison)}
\usage{
\method{anova}{lm_robust}(...)
}
\arguments{
\item{...}{lm_robust objects containing the result from the lm_robust function.}
}
\value{
Outputs the model comparison table (as a data.frame). If not statistically significant,
the simpler model may be preferred.
}
\description{
Computes a deviance table for model comparisons using lm_robust objects.
}
\section{Warning}{

This function requires that the model objects be entered from the simplest
to the more complex model (reduced model first, then full model). The function
requires that the models be nested within each other.
}

\examples{
\dontrun{
library(estimatr)
m1 <- lm_robust(mpg ~ wt, data = mtcars)
m2 <- lm_robust(mpg ~ wt + am, data = mtcars)
m3 <- lm_robust(mpg ~ wt + am + vs, data = mtcars)

anova(m1, m2, m3)
}

}
