% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_jk2_rw.R
\name{build_jk2_rw}
\alias{build_jk2_rw}
\title{build_jk2_rw (Create a matrix of replicate weights for JK2)}
\usage{
build_jk2_rw(zone, rep, weight)
}
\arguments{
\item{zone}{The replicate zone.}

\item{rep}{The replicate member (0 or 1)}

\item{weight}{The weight at level 1.}
}
\value{
Outputs a matrix that can be used in svrepdesign.
}
\description{
Create a matrix of replicate weights.
}
\examples{
\dontrun{
rw <- build_jk2_rw(df2$zone, df2$member, df2$totwgt)
}
}
