% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr1.R
\name{cr1}
\alias{cr1}
\title{cr1}
\usage{
cr1(x, dat = NULL, cluster, ...)
}
\arguments{
\item{x}{An lm or glm object.}

\item{dat}{Original dataset.}

\item{cluster}{The clustering variable (in quotes).}

\item{...}{Extra arguments.}
}
\value{
Returns cluster robust standard errors.
}
\description{
Returns CR1 standard errors.
}
\examples{
# library(sandwich)
g1 <- lm(mpg ~ wt + am, data = mtcars)
cr1(g1, dat = mtcars, cluster = 'cyl')
# sqrt(diag(vcovCL(g1, ~cyl, type = 'HC1')))
g2 <- glm(vs ~ wt + am, data = mtcars, family = binomial)
cr1(g2, cluster = 'cyl')
# sqrt(diag(vcovCL(g2, ~cyl, type = 'HC1')))

}
