% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr3.R
\name{cr3}
\alias{cr3}
\title{cr3}
\usage{
cr3(model, dat = NULL, cluster, df = Inf, adj = TRUE, seonly = TRUE)
}
\arguments{
\item{model}{An lm object.}

\item{dat}{Original dataset.}

\item{cluster}{The clustering variable (in quotes).}

\item{df}{Specify degrees of freedom.}

\item{adj}{Includes few cluster adjustments.}

\item{seonly}{Show the standard error only.}
}
\value{
Returns cluster robust standard errors (seonly = TRUE) or
complete regression output.
}
\description{
Returns CR3 (jackknifed) standard errors.
}
\examples{
g1 <- lm(mpg ~ wt + am, data = mtcars)
cr3(g1, dat = mtcars, cluster = 'cyl')

}
