% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_jk2_zones.R
\name{create_jk2_zones}
\alias{create_jk2_zones}
\title{create_jk2_zones (Create a jacknife zones and replicates)
Create the zones and replicate member prior to making weights}
\usage{
create_jk2_zones(data, psu_id, sort_var, stratum = NULL, seed = NULL)
}
\arguments{
\item{data}{dataset.}

\item{psu_id}{ID of the primary sampling unit.}

\item{sort_var}{The measure of size}

\item{stratum}{The stratum (if present)}

\item{seed}{For replicability.}
}
\value{
Outputs a new dataset with the zone and member number.
Dataset is sorted by MOS. PSUs are paired and then randomly split.
One PSU gets its weight multiplied by two, the other multiplied by zero.
If there is an odd number of units, the last PSU is split in two to form
a pseudo strata.
}
\description{
create_jk2_zones (Create a jacknife zones and replicates)
Create the zones and replicate member prior to making weights
}
\examples{
\dontrun{
rw <- create_jk2_zones(df2, id, mos, stype, seed = 123)
}
}
