% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{ctab}
\alias{ctab}
\title{ctab (Create cross tables)}
\usage{
ctab(data, row_var, col_var, percent_type = "row", show_na = TRUE)
}
\arguments{
\item{data}{The dataframe.}

\item{row_var}{Variable on the y axis.}

\item{col_var}{Variable on the x axis.}

\item{percent_type}{Show percent by "row", "col", or "all".}

\item{show_na}{Whether to show missing values or not.}
}
\value{
Outputs a 'tabyl' / 'data.frame' object.
}
\description{
Creates cross tables showing the frequencies and percentages.
}
\examples{
ctab(mtcars, am, vs)

}
