% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{cv (Coefficient of Variation)}
\usage{
cv(cluster)
}
\arguments{
\item{cluster}{The clustering/grouping variable of interest.}
}
\value{
Outputs the coefficient of variation which represents the degree of
cluster size imbalance (0 = completely balanced).
}
\description{
Computes the coefficient of variation.
}
\examples{
cv(mtcars$cyl)

}
