% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msite.R
\docType{data}
\name{msite}
\alias{msite}
\title{msite (multisite [clustered] data)}
\format{
A data frame with 200 rows and 3 variables:
\describe{
  \item{y}{Outcome of interest.}
  \item{tr}{Treatment assignment dummy code (1 = treatment, 0 = control)}
  \item{gr}{Group, the clustering variable.}
}
}
\usage{
msite
}
\description{
Simulated data for a multisite randomized controlled trial. In all, 10 sites of 20 participants in each site.
}
\keyword{datasets}
