% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2mm.R
\name{r2mm}
\alias{r2mm}
\title{r2mm (Computes r2 for multilevel models)}
\usage{
r2mm(obj)
}
\arguments{
\item{obj}{Output using lmer (merMod)}
}
\value{
A data frame
}
\description{
Computes the r2 for a multilevel model. Provides marginal and conditional r2.
}
\examples{
library(lme4)
data(pkrdd)
m1 <- lmer(abcs ~ avar * takeup + female + (1 | tr), data = pkrdd)
r2mm(m1)

}
