% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_glmm.R
\name{robust_glmm}
\alias{robust_glmm}
\title{robust_glmm (Robust standard errors for glmer/merMod objects)}
\usage{
robust_glmm(model)
}
\arguments{
\item{model}{The glmer model object.}
}
\value{
Outputs a data.frame of results comparing model-based standard errors compared to robust standard errors. diag signifies how much larger the robust vs MBSE is.
}
\description{
Create a matrix of replicate weights.
}
\examples{
\dontrun{
library(lme4)
data(suspend, package = 'MLMusingR')
m1 <- glmer(sus ~ gpa + frpmp + male + (1|school), data = suspend, family = binomial)
robust_glmm(m1)
}

}
