% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1iv.R
\docType{data}
\name{tab1iv}
\alias{tab1iv}
\title{tab1iv (instrumental variable example: Journal of Experimental Education)}
\format{
A data frame with 20 rows and 3 variables:
\describe{
  \item{z}{Treatment assignment dummy code (1 = treatment, 0 = control).}
  \item{t}{Treatment takeup dummy code (1 = took the treatment; 0 = did not take treatment).}
  \item{y}{The outcome of interest.}
}
}
\usage{
tab1iv
}
\description{
Tiny dataset to demonstrate instrumental variable estimation.
}
\keyword{datasets}
