% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildcb.R
\name{wildcb}
\alias{wildcb}
\title{wildcb}
\usage{
wildcb(fml, cluster, dat, B = 999, seed = 0, cr3 = FALSE, param = NULL)
}
\arguments{
\item{fml}{A regression formula.}

\item{cluster}{The clustering variable (in quotes).}

\item{dat}{Original dataset.}

\item{B}{Number of bootstrap replicates.}

\item{seed}{A seed for replicability.}

\item{cr3}{Whether to use the CR3 standard error correction. Only available for unrestricted models.}

\item{param}{Specify the coefficient of interest for a restricted model.}
}
\value{
Regression results.
}
\description{
Fits a wild cluster bootstrapped for an lm model.
}
\examples{
g1 <- wildcb(mpg ~ wt + am, cluster = 'cyl', dat = mtcars)

}
