% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb.R
\name{cb}
\alias{cb}
\title{cb}
\usage{
cb(fml, cluster, dat, B = 1000, group = NULL, interval = "bca", seed = NULL)
}
\arguments{
\item{fml}{A regression formula.}

\item{cluster}{The clustering variable (in quotes).}

\item{dat}{Original dataset.}

\item{B}{Number of bootstrap replicates.}

\item{group}{Specify a group (in quotes) to stratify on.}

\item{interval}{Type of confidence interval ('bca' = bias corrected and accelerated or 'percentile').}

\item{seed}{A seed for replicability.}
}
\value{
Regression results.
}
\description{
Fits a cluster bootstrapped linear model.
}
\examples{
g1 <- cb(mpg ~ wt + am, cluster = 'cyl', dat = mtcars)
g1

}
