% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effclust.R
\name{effclust}
\alias{effclust}
\title{effclust (Compute the effective number of clusters)}
\usage{
effclust(mx, cluster)
}
\arguments{
\item{mx}{The model object (lm or glm).}

\item{cluster}{The clustering/grouping variable of interest.}
}
\value{
Outputs a vector of the effective number of clusters per predictor.
}
\description{
Computes the effective number of clusters in the dataset.
}
\examples{
m1 <- lm(mpg ~ wt, mtcars)
effclust(m1, 'cyl')

}
