% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_ps_subclass.R
\name{pool_ps_subclass}
\alias{pool_ps_subclass}
\title{pool_ps_subclass (Pooling function for propensity scores using sub classifications)}
\usage{
pool_ps_subclass(data, outcome, treatment, subclass)
}
\arguments{
\item{data}{The dataset to use.}

\item{outcome}{The outcome of interest.}

\item{treatment}{The treatment variable.}

\item{subclass}{The subclassification variable.}
}
\value{
Outputs the treatment effect.
}
\description{
Computes the estimand of interest using subclassifications with propensity scores.
}
